requirements.txt
google-generativeai
gradio
pillow
gtts
